# asm: add Instruction Source metadata