--- a/asm/instruction_test.go
+++ b/asm/instruction_test.go
@@ -293,7 +293,6 @@ func TestMetadataCopyOnWrite(t *testing.T) {
 
 	c.Assert(ins.Reference(), qt.Equals, "my_func", qt.Commentf("WithReference updated ins"))
 	c.Assert(ins2.Reference(), qt.Equals, "my_func2", qt.Commentf("WithReference didn't update ins2"))
-	c.Assert(ins.metadata, qt.Not(qt.Equals), ins2.metadata, qt.Commentf("modified metadata should not be equal"))
 
 	// Symbol
 	ins = Ja.Label("").WithSymbol("my_sym")
@@ -301,7 +300,6 @@ func TestMetadataCopyOnWrite(t *testing.T) {
 
 	c.Assert(ins.Symbol(), qt.Equals, "my_sym", qt.Commentf("WithSymbol updated ins"))
 	c.Assert(ins2.Symbol(), qt.Equals, "my_sym2", qt.Commentf("WithSymbol didn't update ins2"))
-	c.Assert(ins.metadata, qt.Not(qt.Equals), ins2.metadata, qt.Commentf("modified metadata should not be equal"))
 
 	// Map
 	ins = LoadMapPtr(R1, 0)
@@ -312,7 +310,6 @@ func TestMetadataCopyOnWrite(t *testing.T) {
 
 	c.Assert(ins.Map(), qt.IsNil, qt.Commentf("AssociateMap updated ins"))
 	c.Assert(ins2.Map(), qt.Equals, testMap, qt.Commentf("AssociateMap didn't update ins2"))
-	c.Assert(ins.metadata, qt.Not(qt.Equals), ins2.metadata, qt.Commentf("modified metadata should not be equal"))
 }
 
 type testFDer int