asm: add Instruction Source metadata

    This commit adds Source metadata to Instruction, which is data about
    the origin of a series of Instructions. A Source can be any fmt.Stringer.

    Instructions.Format will now print the source information in-line with the
    Instruction on which it is set. This results in instructions being annotated
    with original source code when loaded from an ELF file, or custom annotations
    if fmt.Stringers are ssigned using Instruction.WithSource.

asm: use sparse metadata

    Replace the struct based metadata with a linked list. This reduces the amount of memory
    needed when the metadata is sparse, aka when there are more different kinds of metadata.
    This also allows packages to have their own private metadata, which we'll use to store
    BTF ext_infos.

    My initial implementation used a map[interface{}]interface{} to store metadata, but it
    turns out that a linked list is more efficient, assuming that we don't have large amounts
    of metadata on a single instruction.

        name                     old time/op    new time/op    delta
        Metadata/set-4             23.1ns ± 1%    14.6ns ± 0%  -36.73%  (p=0.029 n=4+4)
        Metadata/set_existing-4    15.3ns ± 0%     6.8ns ± 0%  -55.19%  (p=0.029 n=4+4)
        Metadata/get_miss-4        8.09ns ± 0%    2.81ns ± 0%  -65.33%  (p=0.029 n=4+4)

        name                     old alloc/op   new alloc/op   delta
        Metadata/set-4              8.00B ± 0%     8.00B ± 0%     ~     (all equal)
        Metadata/set_existing-4     0.00B          0.00B          ~     (all equal)
        Metadata/get_miss-4         0.00B          0.00B          ~     (all equal)

        name                     old allocs/op  new allocs/op  delta
        Metadata/set-4               1.00 ± 0%      1.00 ± 0%     ~     (all equal)
        Metadata/set_existing-4      0.00           0.00          ~     (all equal)
        Metadata/get_miss-4          0.00           0.00          ~     (all equal)

    I think the reason for this is that small maps are kind of slow compared to
    traversing a linked list by brute force.

elf: add simple benchmark