@lmb 2022-04-15
> I'm wondering if we could fold the value == nil case into m.remove() so we can avoid the fallthrough behaviour and make it more concise.

I looked into this, and it's not more straight forward. I think of `Set` in two stages: first, obtain a `head` that doesn't contain an element with `key` in it. Second, depending on `value`, prepend an element or use the new head directly.