@lmb 2022-04-16
I've rejigged `Set` once more, I think it's now as clear as it'll get. Let me know what you think!